requirements.txt
langgraph>=0.2
langchain-core>=0.3
langchain-openai>=0.2
langchain-google-genai>=2.0
pydantic>=2.0
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
streamlit>=1.36

# Optional — only needed for the Jupyter debug UI (debug_ui.py / SotaGraphUI)
# ipywidgets>=8.0
